move config file to util
